+++ example/lib/pages/normal_player_page.dart
@@ -37,9 +37,14 @@ class _NormalPlayerPageState extends State<NormalPlayerPage> {
       body: Column(
         children: [
           const SizedBox(height: 8),
-          AspectRatio(
-            aspectRatio: 16 / 9,
-            child: BetterPlayer(controller: _betterPlayerController),
+          BetterPlayerMultipleGestureDetector(
+            child: AspectRatio(
+              aspectRatio: 16 / 9,
+              child: BetterPlayer(controller: _betterPlayerController),
+            ),
+            onTap: () {
+              print("Tap!");
+            },
           ),
         ],
       ),