+++ lib/src/core/better_player_controller.dart
@@ -473,7 +473,8 @@ class BetterPlayerController {
       case BetterPlayerDataSourceType.file:
         final file = File(betterPlayerDataSource.url);
         if (!file.existsSync()) {
-          print("File ${file.path} doesn't exists. This may be because "
+          BetterPlayerUtils.log(
+              "File ${file.path} doesn't exists. This may be because "
               "you're acessing file from native path and Flutter doesn't "
               "recognize this path.");
         }