+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 * Fixed audio track selection issue on iOS/Android.
 * Fixed issue where speed which couldn't be applied on iOS was saved in player state.
 * Added support for D-pad navigation using a Android TV remote control (by https://github.com/danielz-nenda)
+* Added `BetterPlayerMultipleGestureDetector` to handle problems with gesture detection
 * Expose getter for `eventListeners` in `BetterPlayerController` (by https://github.com/Letalus)
+* Updated documentation
+* Updated dependencies
 
 ## 0.0.70
 * Fixed file data source exception. Right now user will be only warned.